+++ src/conversions.hpp
@@ -1,0 +1,6 @@
+#include "Hy3Node.hpp"
+
+Axis getAxis(Hy3GroupLayout);
+Axis getAxis(ShiftDirection);
+SearchDirection getSearchDirection(ShiftDirection);
+char directionToChar(ShiftDirection);