+++ src/main.cpp
@@ -38,6 +38,7 @@ APICALL EXPORT PLUGIN_DESCRIPTION_INFO PLUGIN_INIT(HANDLE handle) {
 	CONF("node_collapse_policy", INT, 2);
 	CONF("group_inset", INT, 10);
 	CONF("tab_first_window", INT, 0);
+	CONF("kbd_shift_delta", INT, 20);
 
 	// tabs
 	CONF("tabs:height", INT, 15);